Added order tests using new enhanced humble object
Brian created an AWESOME new humble object that makes writing tests, a
JOY!

Thanks Brian!!!

Jira ticket: CAMS-163
