+++ backend/functions/lib/controllers/orders/orders.controller.test.ts
@@ -2,6 +2,7 @@ import { OrdersController } from './orders.controller';
 import { createMockApplicationContext } from '../../testing/testing-utilities';
 import { ORDERS } from '../../testing/mock-data/orders.mock';
 import { ApplicationContext } from '../../adapters/types/basic';
+import { HumbleQuery } from '../../testing/mock.cosmos-client-humble';
 
 describe('orders controller tests', () => {
   let applicationContext: ApplicationContext;
@@ -11,9 +12,13 @@ describe('orders controller tests', () => {
   });
 
   test('should return all orders', async () => {
+    const mockRead = jest.spyOn(HumbleQuery.prototype, 'fetchAll').mockReturnValue({
+      resources: ORDERS,
+    });
     const controller = new OrdersController(applicationContext);
     const result = await controller.getOrders(applicationContext);
     expect(result.success).toBeTruthy();
     expect(result['body']).toEqual(ORDERS);
+    expect(mockRead).toHaveBeenCalled();
   });
 });
