+++ backend/functions/lib/use-cases/orders/orders.test.ts
@@ -1,6 +1,7 @@
 import { createMockApplicationContext } from '../../testing/testing-utilities';
 import { OrdersUseCase } from './orders';
 import { ORDERS } from '../../testing/mock-data/orders.mock';
+import { HumbleQuery } from '../../testing/mock.cosmos-client-humble';
 import { getOrdersGateway, getOrdersRepository, getRuntimeStateRepository } from '../../factory';
 
 describe('Orders use case', () => {
@@ -10,8 +11,12 @@ describe('Orders use case', () => {
     const runtimeStateRepo = getRuntimeStateRepository(mockContext);
     const ordersRepo = getOrdersRepository(mockContext);
     const useCase = new OrdersUseCase(ordersRepo, ordersGateway, runtimeStateRepo);
+    const mockRead = jest.spyOn(HumbleQuery.prototype, 'fetchAll').mockReturnValue({
+      resources: ORDERS,
+    });
     const result = await useCase.getOrders(mockContext);
     expect(result).toEqual(ORDERS);
+    expect(mockRead).toHaveBeenCalled();
   });
   test('should return a list of orders from the gateway to write to the repo', () => {});
 });
